requirements.txt
pyspark==3.1.2
click==8.0.1
warcio==1.7.4
beautifulsoup4==4.9.3